second ver

+++ main.py
@@ -19,8 +19,8 @@ def calculate(n1, n2, op):
     return result
 
 
-num1 = int(input('Введите первое число: '))
+num1 = float(input('Введите первое число: '))
 op = input('Введите знак (+,-,*,/,^): ')
-num2 = int(input('Введите второе число: '))
+num2 = float(input('Введите второе число: '))
 result = calculate(num1, num2, op)
 print(num1, op, num2, '=', result)
